[Add] AddReviewPhotoCollectionView Cell 생성 (#60)
